requirements.txt
streamlit
pandas
plotly
yfinance
python-dotenv
groq
duckduckgo-search
